#234 Refactor: add return type for method
